Abandon multi-branches (from ALGO)

+++ apps/Problems/src/main/java/algorithmic/iteration/Series.java
@@ -74,7 +74,7 @@ class Series {
 
     sumAllUsingDo();
 
-    sumAllAverage = sum / count;
+    sumAllAverage = (double) sum / count;
 
     return sumAllAverage;
